Nested code
*.codo
Create your file without limit or condition of programming language. Code in nested coding mode, it is simple and it is the best method for creating an automatic mechanism device. Use special characters to nest your natural language code. It is very simple you only need to open the nest and loop it as follows [{("&lt;Code>")}] and or if you want to enter a word function mechanism do it as follows "Word([{(" Code ")}]), Here is a table of contents for special characters 

Symbol "Definition" 
: "Until"
; "With"
, "Result of the above"
& "And"
&& "And and And, but separate element"
@ "Currency"
! "Your exclamation"
!!! "Your continuous exclamation"
¡"Exclamation of the system and or of the operator" 
¡¡¡"Exclamation of the system and or the operator continuously"
$ "String" 
* "Means everything he meets" 
(_) "Wants everything that exists really and virtually" 
_ "Assemble words together mot_mot" 
( "Parenteze opening" 
) "Parentenze closure"
{ "Synthesis opening" 
} "Synthetic closure" 
[ "Opening title and or order"
] "Closing title and or order"
' "Request to read what's next"
"" "Between guilmet, equal the subject and or what is concerned"
,_ "Indicates an assembly and or an assembly, you can also use it to make a line change ex:" bla bla bla ,_ 
blah blah blah""
\n "Indicate to return to the beginning of the text or to return to the word which follows paste, ex \nAllo"
? "Use the query only at the end of a word to return to that word and or do the effect of Sub end Sub (Sub, Sub: A word that    meets its word is automatically a mechanism and closes in loop except nest text because without closing the loop)}], the      code will be pending "
{...} "means: pending" 
< "Means greater than, eg: <MOI> and or *<MOI>(_) "
> "Means greater than, eg: <STILL, ME> and or *<STILL, ME>(_)"
(*) "Means absolutely any code and or file and or anything saved on your hardware device"
^ "Means, title and or power" 
.end "Wants to format or erase, short Lord" 
/ "Wants a sequel and or what follows" 
\ "Means end and where it is the last element" 
{|} "You can use it to get an answer, ex: is that the user clicked {YES | NO}" 
{...} "The three points without the summaries magically means the conversion of one or more elements, ex" again without the          synthe "Hello{...}Bye", in short Hello becomes Bye """
~ "Wants OPERATOR and or title" 
M-A_R "Code to enter when you honestly want to get something" 
Codo "Code to enter to reinforce your code and make it respected asks you, in short no matter the other what is essential is 
      what you have entered and if someone else uses it to contradict you respect for the code of others prevails "
,,, "etc"      
  
  Please note that you have to be very careful when entering a single character and or a character sequence because it might mean something else in man's best friend THE DICTIONARY, in short ex: b = Bit and abc = Know the ABC of a job and or a profession ", For the rest you have carte blanche with the file extension [Codo] and or ^ Codo and or Codo who is our new great friend !! ! _! _ !!!